testing/web-platform/tests/encoding-detection/lt-windows-1257.tentative.html

<!doctype html>
<title>lt windows-1257</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>Tai simboli� kodavimo testas. Kai kurioms kalboms, naudojan�ioms lotyni�kus ra�menis, mums reikia daugiau informacijos, kad gal�tume priimti sprendim�.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1257", 'Expected windows-1257');
    }, "Check detection result");
      done();
};
</script>
